bflibrary: Processing window events while sleeping
This avoids the app not responding.
Also some comment and style fixes.

+++ bflibrary/include/bffile.h
@@ -208,7 +208,7 @@ long LbFileLength(const char *fname);
  *  TbFileFind struct which is then used for LbFileFindNext() and LbFileFindEnd() calls.
  */
 TbResult LbFileFindFirst(const char *filespec, TbFileFind *ffind,
-  unsigned int attributes);
+  ulong attributes);
 
 /** Continues listing directory entries.
  *
